# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GenerateImageInputSize = typeof GenerateImageInputSize[keyof typeof GenerateImageInputSize];


export const GenerateImageInputSize = {
  '1024x1024': '1024x1024',
  '1536x1024': '1536x1024',
  '1024x1536': '1024x1536',
} as const;

export type GenerateImageInputQuality = typeof GenerateImageInputQuality[keyof typeof GenerateImageInputQuality];


export const GenerateImageInputQuality = {
  standard: 'standard',
  high: 'high',
} as const;

export interface GenerateImageInput {
  /** @minLength 1 */
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  style?: string | null;
  size?: GenerateImageInputSize;
  quality?: GenerateImageInputQuality;
  /** @nullable */
  seed?: number | null;
}

export interface ImageUpdate {
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  title?: string | null;
}

export interface VariationInput {
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  style?: string | null;
}

export interface GeneratedImage {
  id: number;
  /** @nullable */
  userId?: string | null;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  title?: string | null;
  imageUrl: string;
  size: string;
  /** @nullable */
  style: string | null;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  seed?: number | null;
  isFavorite: boolean;
  generationTimeMs: number;
  createdAt: string;
}

export interface ImageListResponse {
  images: GeneratedImage[];
  total: number;
  page: number;
  limit: number;
}

export interface Collection {
  id: number;
  /** @nullable */
  userId?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  imageCount: number;
  createdAt: string;
}

export interface CollectionDetail {
  id: number;
  /** @nullable */
  userId?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  imageCount: number;
  images: GeneratedImage[];
  createdAt: string;
}

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface CollectionUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface CollectionImageInput {
  imageId: number;
}

export interface AnalyticsSummary {
  totalImages: number;
  totalUsers: number;
  totalCollections: number;
  totalFavorites: number;
  imagesThisWeek: number;
  imagesThisMonth: number;
}

export interface DailyStats {
  date: string;
  count: number;
}

export interface StyleStat {
  style: string;
  count: number;
}

export type ListImagesParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
style?: string | null;
/**
 * @nullable
 */
favorite?: boolean | null;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type ListRecentImagesParams = {
limit?: number;
};

export type GetDailyAnalyticsParams = {
days?: number;
};