# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Text-to-Image Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List user's collections
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a collection with its images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection details with images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}/images:
    post:
      operationId: addImageToCollection
      tags: [collections]
      summary: Add an image to a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionImageInput"
      responses:
        "200":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}/images/{imageId}:
    delete:
      operationId: removeImageFromCollection
      tags: [collections]
      summary: Remove an image from a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    GeneratedImage:
      type: object
      required: [id, prompt, imageUrl, size, style, isFavorite, generationTimeMs, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        prompt:
          type: string
        negativePrompt:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        imageUrl:
          type: string
        size:
          type: string
        style:
          type: ["string", "null"]
        quality:
          type: ["string", "null"]
        seed:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        generationTimeMs:
          type: integer
        createdAt:
          type: string
...
    Collection:
      type: object
      required: [id, name, imageCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        imageCount:
          type: integer
        createdAt:
          type: string
    CollectionDetail:
      type: object
      required: [id, name, images, imageCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        imageCount:
          type: integer
        images:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedImage"
        createdAt:
          type: string
    CollectionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
    CollectionUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    CollectionImageInput:
      type: object
      required: [imageId]
      properties:
        imageId:
          type: integer
...