# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  /** @nullable */
  userId?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  imageCount: number;
  createdAt: string;
}