# lib/api-zod/src/generated/types/collectionDetail.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedImage } from './generatedImage';

export interface CollectionDetail {
  id: number;
  /** @nullable */
  userId?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  imageCount: number;
  images: GeneratedImage[];
  createdAt: string;
}