# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
}