# lib/api-zod/src/generated/types/collectionUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
}