# lib/api-zod/src/generated/types/generatedImage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedImage {
  id: number;
  /** @nullable */
  userId?: string | null;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  title?: string | null;
  imageUrl: string;
  size: string;
  /** @nullable */
  style: string | null;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  seed?: number | null;
  isFavorite: boolean;
  generationTimeMs: number;
  createdAt: string;
}